test(renderer-jsx): unify markdown component assertions on renderToText helper
Extract the `firstSourceText` lookup that was duplicated across the six
markdown component tests into a shared `renderToText(children)` helper in
`_testUtils.tsx`. Each test now reads as a single
`expect(await renderToText(<Component … />)).toBe(…)` line, so all six
files share the same assertion shape.
